requirements.txt
alembic==1.4.2
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycryptodome==3.3.1
Pygments==2.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
zipp==3.1.0
